requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
art==5.6
async-timeout==4.0.2
attrs==21.4.0
charset-normalizer==2.0.12
demjson==2.2.4
frozenlist==1.3.0
idna==3.3
multidict==6.0.2
py-cord==2.0.0rc1
python-dotenv==0.20.0
yarl==1.7.2
